chore(lunox-prisma): change bundler to tsup
