+++ packages/lunox-prisma/package.json
@@ -10,8 +10,7 @@
   },
   "files": [
     "dist/*",
-    "contracts.d.ts",
-    "src/contracts/index.d.ts"
+    "contracts.d.ts"
   ],
   "scripts": {
     "build": "lunox prod",
@@ -32,13 +31,11 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@lunoxjs/build": "2.0.0-next.33",
     "@lunoxjs/core": "2.0.0-next.33",
     "@lunoxjs/test": "2.0.0-next.33",
     "@prisma/client": "^5.6.0",
     "@types/supertest": "^2.0.16",
     "prisma": "^5.6.0",
-    "rollup": "^4.5.2",
     "typescript": "^5.3.2",
     "vitest": "^0.34.6"
   },
