--- packages/lunox-prisma/rollup.config.mjs
@@ -1,14 +1,0 @@
-import { bundleTs } from "@lunoxjs/build";
-export default [
-  ...bundleTs(["src/index.ts", "src/contracts/index.ts"], {
-    declaration: process.env.NODE_ENV == "production",
-    rollupConfig: {
-      external: ["@prisma/client", "prisma"],
-    },
-    dtsConfig: {
-      compilerOptions: {
-        skipLibCheck: true,
-      },
-    },
-  }),
-];
