--- packages/lunox-prisma/src/contracts/index.d.ts
@@ -1,11 +1,0 @@
-import type { Prisma, PrismaClient } from "@prisma/client";
-import type Seeder from "../Seeder";
-
-export interface DatabaseConfig {
-  url: string;
-  client: typeof PrismaClient;
-  log?: Prisma.PrismaClientOptions["log"];
-  errorFormat?: Prisma.PrismaClientOptions["errorFormat"];
-  baseSeeder?: typeof Seeder;
-}
-export declare class PrismaClientLinked extends PrismaClient { }
