+++ packages/lunox-prisma/tsconfig.json
@@ -1,7 +1,6 @@
 {
-  "extends": "@lunoxjs/build/tsconfig.json",
+  "extends": "../lunox-build/tsconfig.json",
   "compilerOptions": {
     "typeRoots": ["node_modules/@types"]
-  },
-  "exclude": ["@lunoxjs/core/dist"]
+  }
 }
